+++ src/aui-widget/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-2021](https://issues.liferay.com/browse/AUI-2021) AUI-Popover should choose the best position according to its constrain element
 
 ## [3.0.1](https://github.com/liferay/alloy-ui/releases/tag/3.0.1)